add tests + merge components
