+++ packages/ui/src/components/_ui/table-pagination.stories.tsx
@@ -1,15 +1,16 @@
-import { Column, TableAuto } from './table-auto'
+import { TablePagination } from './table-pagination'
 import { Meta, Story } from '@storybook/vue3'
 import { markRaw } from 'vue'
 import TitreNom from '../_common/titre-nom.vue'
 import { Domaine } from '../_common/domaine'
 import TitreTypeTypeNom from '../_common/titre-type-type-nom.vue'
 import { Statut } from '../_common/statut'
-import { TableRow } from './table'
+import { Column, TableRow } from './table'
+import { action } from '@storybook/addon-actions'
 
 const meta: Meta = {
   title: 'Components/UI/Table',
-  component: TableAuto
+  component: TablePagination
 }
 export default meta
 
@@ -39,7 +40,7 @@ const columns: Column[] = [
   }
 ]
 
-const rows: TableRow[] = [0, 1, 2, 3].map(row => {
+const rows: TableRow[] = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11].map(row => {
   return {
     id: `elementId${row}`,
     link: {
@@ -84,13 +85,14 @@ const rows: TableRow[] = [0, 1, 2, 3].map(row => {
   }
 })
 
-export const TableAutoSimple: Story = () => (
-  <TableAuto rows={rows} columns={columns} />
-)
-export const TableAutoSortedByStatusAsc: Story = () => (
-  <TableAuto
+const paramsUpdate = action('paramsUpdate')
+export const PaginationSimple: Story = () => (
+  <TablePagination
     rows={rows}
     columns={columns}
-    initialSort={{ column: 'statut', order: 'desc' }}
+    total={200}
+    paramsUpdate={paramsUpdate}
+    pagination={true}
+    range={10}
   />
 )
