+++ packages/ui/src/components/_ui/table-pagination.tsx
@@ -1,7 +1,10 @@
-import { computed, defineComponent } from 'vue'
-import { Column, Table, TableRow, TableSortEvent } from './newTable'
-import { Pagination } from './pagination'
-import { Ranges } from './ranges'
+import { computed, defineComponent, FunctionalComponent, ref } from 'vue'
+import { Column, Table, TableRow, TableSortEvent } from './table'
+import { isEventWithTarget } from '@/utils/vue-tsx-utils'
+import Accordion from './accordion.vue'
+
+const ranges = [10, 50, 200, 500] as const
+type Range = (typeof ranges)[number]
 
 export interface Params {
   page?: number
@@ -11,7 +14,7 @@ interface Props {
   columns: readonly Column[]
   rows: TableRow[]
   total: number
-  range?: number
+  range?: Range
   page?: number
   column?: string
   order?: 'asc' | 'desc'
@@ -19,7 +22,6 @@ interface Props {
   paramsUpdate: (params: Params | TableSortEvent) => void
 }
 
-// FIXME TEST
 export const TablePagination = defineComponent<Props>({
   props: [
     'columns',
@@ -91,11 +93,7 @@ export const TablePagination = defineComponent<Props>({
             </div>
             <div class="desktop-blob-1-4">
               {props.total > 10 ? (
-                <Ranges
-                  ranges={[10, 50, 200, 500]}
-                  range={range.value}
-                  rangeUpdate={rangeUpdate}
-                />
+                <Ranges range={range.value} rangeUpdate={rangeUpdate} />
               ) : null}
             </div>
           </div>
@@ -104,3 +102,155 @@ export const TablePagination = defineComponent<Props>({
     )
   }
 })
+
+interface RangeProps {
+  range: Range
+  rangeUpdate: (range: Range) => void
+}
+
+const isRange = (range: number): range is Range => ranges.includes(range)
+const Ranges = defineComponent<RangeProps>({
+  props: ['range', 'rangeUpdate'] as unknown as undefined,
+  setup(props) {
+    const opened = ref(false)
+
+    const toggle = () => {
+      opened.value = !opened.value
+    }
+
+    return () => (
+      <Accordion
+        class="mb"
+        opened={opened.value}
+        slotDefault={true}
+        onToggle={toggle}
+      >
+        {{
+          title: () => <span> Éléments </span>,
+          default: () => (
+            <ul class="list-sans mt-m px-m">
+              {ranges.map(r => (
+                <li key={r}>
+                  <label>
+                    <input
+                      checked={r === props.range}
+                      value={r}
+                      type="radio"
+                      class="mr-s"
+                      onChange={e => {
+                        if (isEventWithTarget(e)) {
+                          const value = Number(e.target.value)
+                          if (isRange(value)) {
+                            props.rangeUpdate(value)
+                          }
+                        }
+                      }}
+                    />
+                    {r}
+                  </label>
+                </li>
+              ))}
+            </ul>
+          )
+        }}
+      </Accordion>
+    )
+  }
+})
+
+interface PaginationProps {
+  total: number
+  active: number
+  visibles: number
+  pageChange: (page: number) => void
+}
+
+const Pagination: FunctionalComponent<PaginationProps> = props => {
+  const total = props.total ?? 2
+  const active = props.active ?? 1
+  const visibles = props.visibles ?? 1
+
+  const delta = () => {
+    return Math.round((visibles - 1) / 2)
+  }
+  const pages = () => {
+    let filter
+    if (active <= delta()) {
+      filter = (n: number) => n <= delta() * 2 + 1
+    } else if (active >= total - delta()) {
+      filter = (n: number) => n >= total - delta() * 2
+    } else {
+      filter = (n: number) => n >= active - delta() && n <= active + delta()
+    }
+    return Array.from(Array(total).keys())
+      .map(n => n + 1)
+      .filter(filter)
+  }
+
+  if (total <= 1) {
+    return null
+  }
+  return (
+    <ul class="list-inline">
+      <li class="mr-xs">
+        <button
+          disabled={active === 1}
+          class="btn-border rnd-xs px-m py-s"
+          onClick={() => props.pageChange(1)}
+        >
+          «
+        </button>
+      </li>
+      <li class="mr-xs">
+        <button
+          disabled={active === 1}
+          class="btn-border rnd-xs px-m py-s"
+          onClick={() => props.pageChange(active - 1)}
+        >
+          ‹
+        </button>
+      </li>
+      {active > delta() + 1 ? (
+        <li class="mr-xs">
+          <div class="px-m py-s">…</div>
+        </li>
+      ) : null}
+
+      {pages().map(page => (
+        <li key={page} class={`mr-xs ${active === page ? 'active' : ''}`}>
+          <button
+            class="btn-border rnd-xs px-m py-s"
+            onClick={() => props.pageChange(page)}
+          >
+            {page}
+          </button>
+        </li>
+      ))}
+
+      {active < total - delta() ? (
+        <li class="mr-xs">
+          <div class="px-m py-s">…</div>
+        </li>
+      ) : null}
+
+      <li class="mr-xs">
+        <button
+          disabled={active === total}
+          class="btn-border rnd-xs px-m py-s"
+          onClick={() => props.pageChange(active + 1)}
+        >
+          ›
+        </button>
+      </li>
+      <li class="mr-xs">
+        <button
+          disabled={active === total}
+          class="btn-border rnd-xs px-m py-s"
+          onClick={() => props.pageChange(total)}
+        >
+          »
+        </button>
+      </li>
+    </ul>
+  )
+}
