+++ packages/ui/src/components/_ui/table.tsx
@@ -59,7 +59,6 @@ interface Props {
   order: 'asc' | 'desc'
 }
 
-// FIXME TEST
 export const Table = defineComponent<Props>({
   props: [
     'columns',
