--- packages/ui/src/components/_ui/table.vue
@@ -1,110 +1,0 @@
-<template>
-  <div>
-    <div class="overflow-scroll-x mb">
-      <div class="table">
-        <div class="tr">
-          <div
-            v-for="col in columns"
-            :key="col.id"
-            class="th nowrap"
-            :class="col.class"
-            @click="sort(col.id)"
-          >
-            <button
-              class="btn-menu full-x p-0"
-              :class="{ disabled: col.noSort }"
-            >
-              {{ col.name || (column === col.id ? '' : '–') }}
-              <Icon
-                v-if="!col.noSort && column === col.id"
-                class="right"
-                size="M"
-                :name="order === 'asc' ? 'chevron-bas' : 'chevron-haut'"
-              />
-            </button>
-          </div>
-        </div>
-
-        <router-link
-          v-for="row in rows"
-          :key="row.id"
-          :to="row.link"
-          class="tr tr-link text-decoration-none"
-        >
-          <div
-            v-for="col in columns"
-            :key="col.id"
-            class="td"
-            :class="col.class"
-          >
-            <component
-              :is="row.columns[col.id].component"
-              v-if="
-                row.columns[col.id] &&
-                row.columns[col.id].component &&
-                row.columns[col.id].slot
-              "
-              v-bind="row.columns[col.id].props"
-              :class="row.columns[col.id].class"
-              >{{ row.columns[col.id].value }}</component
-            >
-            <component
-              :is="row.columns[col.id].component"
-              v-else-if="row.columns[col.id] && row.columns[col.id].component"
-              v-bind="row.columns[col.id].props"
-              :class="row.columns[col.id].class"
-            />
-            <span
-              v-else-if="row.columns[col.id] && row.columns[col.id].value"
-              :class="row.columns[col.id].class"
-              >{{ row.columns[col.id].value }}</span
-            >
-          </div>
-        </router-link>
-      </div>
-    </div>
-  </div>
-</template>
-
-<script>
-import { Icon } from '@/components/_ui/icon'
-export default {
-  name: 'UiTable',
-  components: { Icon },
-  props: {
-    rows: { type: Array, required: true },
-    columns: { type: Array, required: true },
-    order: { type: String, default: 'asc' },
-    column: { type: String, default: '' }
-  },
-
-  emits: ['params-update'],
-
-  watch: {
-    columns: 'columnInit'
-  },
-
-  methods: {
-    update(params) {
-      this.$emit('params-update', params)
-    },
-
-    sort(colId) {
-      if (!this.columns.find(c => c.id === colId).noSort) {
-        if (this.column === colId) {
-          const order = this.order === 'asc' ? 'desc' : 'asc'
-          this.update({ order, column: this.column })
-        } else {
-          this.update({ column: colId, order: this.order })
-        }
-      }
-    },
-
-    columnInit() {
-      if (this.rows.length && !this.columns.some(c => c.id === this.column)) {
-        this.sort(this.columns[0].id)
-      }
-    }
-  }
-}
-</script>
